feat: Add AuthService interface
