+++ src/main/java/com/synergy/backend/domain/member/exception/ErrorType.java
@@ -11,7 +11,7 @@ public enum ErrorType {
 	_ACCESS_DENIED(403, "접근 권한이 없습니다."),
 	_ADMIN_OR_RECRUITER_NOT_FOUND(404, "해당 관리자나 채용담당자를 찾을 수 없습니다."),
 	_INVALID_AUTH_CODE(400, "유효하지 않은 인증코드 입니다."),
-	_INVALID_PASSWORD(400, "비밀번호가 일치하지 않습니다."),
+	_INVALID_EMAIL_OR_PASSWORD(400, "이메일이 존재하지 않거나 비밀번호가 다릅니다."),
 	_DUPLICATE_EMAIL_EXCEPTION(409, "이미 등록된 이메일입니다.")
 
 	;
