--- src/main/java/com/synergy/backend/domain/member/exception/InvalidPasswordException.java
@@ -1,9 +1,0 @@
-package com.synergy.backend.domain.member.exception;
-
-import com.synergy.backend.global.exception.BaseErrorException;
-
-public class InvalidPasswordException extends BaseErrorException {
-	public InvalidPasswordException() {
-		super(ErrorType._INVALID_PASSWORD.getCode(), ErrorType._INVALID_PASSWORD.getMessage());
-	}
-}
