+++ src/test/java/com/synergy/backend/domain/member/service/AuthServiceImplTest.java
@@ -1,0 +1,12 @@
+package com.synergy.backend.domain.member.service;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.BeforeEach;
+
+class AuthServiceImplTest {
+
+	@BeforeEach
+	void setUp() {
+	}
+}
